requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
binance-connector==3.12.0
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
dateparser==1.2.1
debugpy==1.8.11
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.4.3
nest_asyncio==1.6.0
numpy==2.2.4
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==25.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodome==3.22.0
Pygments==2.19.1
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.11.1
pytz==2025.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
stack_data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
yarl==1.19.0
zipp==3.21.0
